Move storage to the bot
