+++ src/bot/storage/json.ts
@@ -1,8 +1,8 @@
 import path from "node:path";
 import fs from "node:fs/promises";
-import { createLogger } from "./../utils/logger.js";
-import type { TransactionRow, TransactionStorage } from "../types.js";
-import { createSaveStats } from "../bot/saveStats.js";
+import { createLogger } from "../../utils/logger.js";
+import type { TransactionRow, TransactionStorage } from "../../types.js";
+import { createSaveStats } from "../saveStats.js";
 
 const logger = createLogger("LocalJsonStorage");
